requirements.txt
# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# Security
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
flask-talisman==1.1.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
PyJWT==2.10.1

# Validation & Sanitization
marshmallow==4.2.0
bleach==6.3.0
email-validator==2.3.0
WTForms==3.2.1

# Rate Limiting
limits==5.6.0
ordered-set==4.1.0
Deprecated==1.3.1
wrapt==2.0.1

# HTTP & Utilities
requests==2.32.5
certifi==2026.1.4
charset-normalizer==3.4.4
urllib3==2.6.3
idna==3.11
dnspython==2.8.0

# Core Dependencies
blinker==1.9.0
click==8.3.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.5
greenlet==3.3.0
typing_extensions==4.15.0

# Other
cffi==2.0.0
pycparser==2.23
webencodings==0.5.1
packaging==25.0